// src/api/generated/model/postDraftRequest.ts
/**
 * Generated by orval v8.0.0-rc.2 🍺
 * Do not edit manually.
 * OpenAPI definition
 * OpenAPI spec version: v0
 */
import type { PostDraftRequestCategory } from './postDraftRequestCategory';

export interface PostDraftRequest {
  title?: string;
  content?: string;
  imageUrl?: string;
  category?: PostDraftRequestCategory;
  tags?: string[];
}